refactor interrupt with no debug into applicationinterrupr
